資料表欄位變動

+++ database/migrations/2025_02_18_024020_create_courses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id(); // 自動遞增的主鍵
             $table->string('title'); // 課程名稱
             $table->text('description')->nullable(); // 課程描述
-            $table->decimal('price', 10, 2)->default(0); // 課程價格
+            $table->integer('price')->default(0); // 課程價格
             $table->unsignedBigInteger('teacher_id'); // 講師（關聯 users）
             $table->unsignedBigInteger('category_id')->nullable(); // 課程分類（關聯 categories）
             $table->string('cover_image')->nullable(); // 課程封面圖片
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->timestamps(); // 自動管理 created_at & updated_at
 
             // 設定外鍵
-            
+
         });
     }
